refactor: Change file's path And Delete legacy files
